Commit shel bnei zonot !!!!

+++ src/Controller/doctorController.Java
@@ -28,7 +28,7 @@ public class doctorController {
 			this.currentModelDoctor = new Doctor (id, phone, name, email, special,password,user_type);
 	}
 	
-	public boolean addNewPatient (String id,int age, int phone, String name, String email, int weight, int height, String gender, String allergies, String subscriptions, String chronic_diseases) {
+	public boolean addNewPatient (String id,int age, long phone, String name, String email, int weight, int height, String gender, String allergies, String subscriptions, String chronic_diseases) {
 		
 		Patient newPatient = new Patient(id, age, phone, name, email, weight, height, gender, allergies, subscriptions, chronic_diseases);
 		if (currentModelDoctor.add(newPatient))
